[ANDROSDK-1091] Move BaseTests to core so D2DIComponent doesn't have to be reinstantiated
