+++ core/src/androidTest/java/org/hisp/dhis/android/core/BaseRealIntegrationTest.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.utils.integration.real;
+package org.hisp.dhis.android.core;
 
 import android.content.Context;
 
@@ -34,16 +34,10 @@ import androidx.test.InstrumentationRegistry;
 
 import com.facebook.stetho.Stetho;
 
-import org.hisp.dhis.android.core.D2;
 import org.hisp.dhis.android.core.arch.call.internal.GenericCallData;
 import org.hisp.dhis.android.core.arch.d2.internal.D2DIComponent;
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
 import org.hisp.dhis.android.core.arch.db.access.internal.DatabaseAdapterFactory;
-import org.hisp.dhis.android.core.arch.storage.internal.AndroidSecureStore;
-import org.hisp.dhis.android.core.arch.storage.internal.Credentials;
-import org.hisp.dhis.android.core.arch.storage.internal.CredentialsSecureStoreImpl;
-import org.hisp.dhis.android.core.arch.storage.internal.ObjectSecureStore;
-import org.hisp.dhis.android.core.arch.storage.internal.SecureStore;
 import org.hisp.dhis.android.core.data.server.RealServerMother;
 import org.hisp.dhis.android.core.resource.internal.ResourceHandler;
 import org.junit.After;
@@ -59,8 +53,6 @@ public abstract class BaseRealIntegrationTest {
 
     protected Date serverDate = new Date();
     protected ResourceHandler resourceHandler;
-    private SecureStore secureStore;
-    protected ObjectSecureStore<Credentials> credentialsSecureStore;
 
     protected String username = RealServerMother.username;
     protected String password = RealServerMother.password;
@@ -72,8 +64,6 @@ public abstract class BaseRealIntegrationTest {
 
         databaseAdapter = DatabaseAdapterFactory.getDatabaseAdapter();
         DatabaseAdapterFactory.createOrOpenDatabase(databaseAdapter, null, context, false);
-        secureStore = new AndroidSecureStore(context);
-        credentialsSecureStore = new CredentialsSecureStoreImpl(secureStore);
         resourceHandler = ResourceHandler.create(databaseAdapter);
         resourceHandler.setServerDate(serverDate);
         Stetho.initializeWithDefaults(context);
@@ -101,7 +91,6 @@ public abstract class BaseRealIntegrationTest {
     }
 
     protected D2DIComponent getD2DIComponent(D2 d2) {
-        return D2DIComponent.create(InstrumentationRegistry.getTargetContext().getApplicationContext(), d2.retrofit(),
-                databaseAdapter, secureStore);
+        return d2.d2DIComponent;
     }
 }
