+++ core/src/androidTest/java/org/hisp/dhis/android/core/MockIntegrationTestObjects.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.utils.integration.mock;
+package org.hisp.dhis.android.core;
 
 import android.content.Context;
 import android.util.Log;
@@ -35,15 +35,12 @@ import androidx.test.InstrumentationRegistry;
 
 import com.facebook.stetho.Stetho;
 
-import org.hisp.dhis.android.core.D2;
-import org.hisp.dhis.android.core.D2Factory;
 import org.hisp.dhis.android.core.arch.d2.internal.D2DIComponent;
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
-import org.hisp.dhis.android.core.arch.storage.internal.InMemorySecureStore;
-import org.hisp.dhis.android.core.arch.storage.internal.SecureStore;
 import org.hisp.dhis.android.core.data.server.Dhis2MockServer;
 import org.hisp.dhis.android.core.period.internal.CalendarProviderFactory;
 import org.hisp.dhis.android.core.resource.internal.ResourceHandler;
+import org.hisp.dhis.android.core.utils.integration.mock.MockIntegrationTestDatabaseContent;
 
 import java.io.IOException;
 import java.util.Date;
@@ -59,7 +56,7 @@ public class MockIntegrationTestObjects {
     public final Dhis2MockServer dhis2MockServer;
     public final MockIntegrationTestDatabaseContent content;
 
-    MockIntegrationTestObjects(MockIntegrationTestDatabaseContent content) throws Exception {
+    public MockIntegrationTestObjects(MockIntegrationTestDatabaseContent content) throws Exception {
         this.content = content;
 
         deleteDatabase();
@@ -73,9 +70,7 @@ public class MockIntegrationTestObjects {
         d2 = D2Factory.forNewDatabase();
 
         databaseAdapter = d2.databaseAdapter();
-        SecureStore secureStore = new InMemorySecureStore();
-
-        d2DIComponent = D2DIComponent.create(context, d2.retrofit(), databaseAdapter, secureStore);
+        d2DIComponent = d2.d2DIComponent;
 
         resourceHandler = ResourceHandler.create(databaseAdapter);
         resourceHandler.setServerDate(serverDate);
