+++ core/src/androidTest/java/org/hisp/dhis/android/core/note/NotePostCallRealIntegrationShould.java
@@ -28,11 +28,11 @@
 
 package org.hisp.dhis.android.core.note;
 
+import org.hisp.dhis.android.core.BaseRealIntegrationTest;
 import org.hisp.dhis.android.core.D2;
 import org.hisp.dhis.android.core.D2Factory;
 import org.hisp.dhis.android.core.data.server.RealServerMother;
 import org.hisp.dhis.android.core.enrollment.Enrollment;
-import org.hisp.dhis.android.core.utils.integration.real.BaseRealIntegrationTest;
 import org.junit.Before;
 
 import java.io.IOException;
